# faster narrow mmt4d ukernels on x86