@bjacob 2024-03-04
@pashu123 , this PR might be a useful example to study. Note: to build and run the self-contained test and benchmark for these ukernels:

```
ninja mmt4d_test mmt4d_benchmark && ./runtime/src/iree/builtins/ukernel/tools/mmt4d_test && ./runtime/src/iree/builtins/ukernel/tools/mmt4d_benchmark
```